fix: ajuste de logica de locacao de produto por tipo de caixa

+++ SmartPacking/Model/BoxModel.cs
@@ -11,6 +11,6 @@ namespace SmartPacking.Model
         public int Width { get; set; }
         public int Length { get; set; }
 
-        public double Volume => Height * Width * Length;
+        public int Volume => Height * Width * Length;
     }
 }
