+++ SmartPacking/Model/ProductModel.cs
@@ -15,7 +15,7 @@ namespace SmartPacking.Model
         public int Width { get; set; }
         [Required]
         public int Length { get; set; }
-        public double Volume => Height * Width * Length;
+        public int Volume => Height * Width * Length;
 
         #region Relacionamentos
         public int OrderModelId { get; set; }
